update slider in setting screen
